Refactor system menu to make it work more like the other menus
